convert GetEnvelopeFromFileHelper from Multilaser to Laser
