+++ src/laser/Laser.H
@@ -19,7 +19,7 @@ class Laser
 {
 public:
 
-    Laser (std::string name, bool laser_from_file);
+    Laser (std::string name);
 
     std::string m_name {""};
     /** The way to initialize a laser (from_file/gaussian/parser)*/
@@ -44,6 +44,27 @@ public:
     amrex::ParserExecutor<3> m_profile_real;
     /** stores the output function of makeFunctionWithParser for profile_imag_str */
     amrex::ParserExecutor<3> m_profile_imag;
+
+    /** if the lasers are initialized from openPMD file */
+    bool m_laser_from_file = false;
+    /** full 3D laser data stored on the host */
+    amrex::FArrayBox m_F_input_file;
+    /** path to input openPMD file */
+    std::string m_input_file_path;
+    /** name of the openPMD species in the file */
+    std::string m_file_envelope_name = "laserEnvelope";
+    /** index of the iteration in the openPMD file */
+    int m_file_num_iteration = 0;
+    /** Geometry of the laser file, 'rt' or 'xyt' */
+    std::string m_file_geometry = "";
+    /** this is to pass the value to multilaser */
+    amrex::Real m_lambda0 {0.};
+    /** \brief Read in a laser from an openPMD file */
+    void GetEnvelopeFromFileHelper ();
+
+    /** \brief Read in a laser from an openPMD file */
+    template<typename input_type>
+    void GetEnvelopeFromFile ();
 };
 
 #endif // LASER_H_
