+++ src/laser/MultiLaser.H
@@ -88,13 +88,6 @@ public:
      */
     void InitSliceEnvelope (const int islice, const int comp);
 
-    /** \brief Read in a laser from an openPMD file */
-    void GetEnvelopeFromFileHelper ();
-
-    /** \brief Read in a laser from an openPMD file */
-    template<typename input_type>
-    void GetEnvelopeFromFile ();
-
     /** \brief Shift 2D slices in zeta
      * \param[in] islice slice index
      */
@@ -198,8 +191,9 @@ private:
 
     bool m_use_laser {false}; /**< whether a laser is used or not */
     /** Laser central wavelength.
-     * he central wavelength influences the solver. As long as all the lasers are on the same grid
-     * (part of MultiLaser), this must be a property of MultiLaser. */
+     * the central wavelength influences the solver. As long as all the lasers are on the same grid
+     * (part of MultiLaser), this must be a property of MultiLaser.
+     * If laser.init_type = 'from_file', lambda0 would be covered by the wavelength in the input_file */
     amrex::Real m_lambda0 {0.};
     amrex::Vector<std::string> m_names {"no_laser"}; /**< name of the laser */
     int m_nlasers; /**< Number of laser pulses */
@@ -218,19 +212,6 @@ private:
     amrex::Box m_slice_box;
     /** interpolation order for laser to field and field to laser operations */
     int m_interp_order = 1;
-
-    /** if the lasers are initialized from openPMD file */
-    bool m_laser_from_file = false;
-    /** full 3D laser data stored on the host */
-    amrex::FArrayBox m_F_input_file;
-    /** path to input openPMD file */
-    std::string m_input_file_path;
-    /** name of the openPMD species in the file */
-    std::string m_file_envelope_name = "laserEnvelope";
-    /** index of the iteration in the openPMD file */
-    int m_file_num_iteration = 0;
-    /** Geometry of the laser file, 'rt' or 'xyt' */
-    std::string m_file_geometry = "";
 
     /** Array of N slices required to compute current slice */
     amrex::MultiFab m_slices;
